Create file for Order page styled components
Migrate styled components to assigned file

Update file imports
