+++ src/components/OrderPage/Order2.tsx
@@ -5,11 +5,10 @@ import { useDispatch } from "react-redux";
 import { useNavigate } from "react-router-dom";
 import { collection, serverTimestamp, doc, setDoc, getCountFromServer } from "firebase/firestore"
 
-import styled from "styled-components";
-import { OrderBody } from "../../app/Utils/StyledComponents/CartComponents";
 import CartItem from "../CartPage/CartItem";
 import CardModal from "./CardModal";
 import FinalOrder from "./FinalOrder";
+import { Address, Order2Body, Order2Container, OrderButton } from "../../app/Utils/StyledComponents/OrderComponents";
 
 import { useAppSelector } from "../../app/Utils/hooks/useAppSelector";
 import { useCartState } from "../../app/Utils/hooks/useCartState";
@@ -151,31 +150,31 @@ const Order2 = (props: order2Props) => {
         </p>
 
         <FinalOrder subTot={subTotal} />
-        <Button
+        <OrderButton
           onClick={() => toggleCardModal()}
         >
           Card
-        </Button>
+        </OrderButton>
         <p>
           Add Gift Card
         </p>
         <p> By continuing, I understand and agree to the General Terms and Conditions of Online Accessories Sales, <span>Terms of Use</span> and <span>Privacy Notice</span>.</p>
         {cardDet.name && cardDet.cvv && cardDet.exp && cardDet.number ?
-          <Button
+          <OrderButton
             className="px-8"
             onClick={(e: BaseSyntheticEvent) => {
               logOrder(e)
             }}
           >
             Place Order
-          </Button>
+          </OrderButton>
         :
-          <Button
+          <OrderButton
             className="px-4"
             disabled
           >
             Place Order
-          </Button>
+          </OrderButton>
           }
         </Order2Container>
     </Order2Body>
@@ -184,113 +183,3 @@ const Order2 = (props: order2Props) => {
 };
 
 export default Order2;
-
-const Order2Body = styled(OrderBody)`
-  >h1{
-    position: relative;
-    bottom: 16px;
-    font-size: 60px;
-    font-weight: 500;
-  }
-  >p{
-    font-size: 19px;
-    letter-spacing: 1.05px;
-  }
-`;
-
-const Order2Container = styled.div`
-  display: flex;
-  >ul{
-    padding: 50px 0;
-    margin-bottom: 70px;
-    border-bottom: 1px solid grey;
-  }
-  >p:first-of-type{
-    font-size: 19px;
-    font-weight: 500;
-    margin-bottom: 34px;
-    color: rgba(30,30,30)
-  }
-  >p:nth-of-type(2){
-    font-size: 18px;
-    font-weight: 350;
-    letter-spacing: 1.04px;
-    margin-bottom: 32px;
-  }
-  >p:nth-of-type(3){
-    font-size: 16px;
-    font-weight: 500;
-    margin-bottom: 30px;
-    letter-spacing: 1.04px;
-    word-spacing: 1.2px;
-    color: rgba(90, 90, 90);
-    >span{
-      color: inherit;
-      border-bottom: 1.5px solid black;
-      :hover{
-        border-bottom: 2px solid black;
-        color: black;
-      }
-    }
-    
-  }
-  @media screen and (min-width: 1280px) {
-
-  }
-  @media screen and (max-width: 1280px) {
-    flex-direction: column;
-  }
-`;
-
-const Address = styled.div`
-  letter-spacing: 1.1px;
-  line-height: 28px;
-  margin-bottom: 27px;
-  >div{
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    >h3{
-      font-size: 20px;
-      font-weight: 350;
-      color: rgba(90, 90, 90);
-      margin-bottom: 6px;
-    }
-    >button{
-      position: relative;
-      right: 26px;
-      font-size: 20px;
-      color: rgba(80, 80, 80);
-      border-bottom: 1.5px solid rgba(80, 80, 80);
-      line-height: 24px;
-      :hover{
-        border-bottom: 3px solid rgba(80, 80, 80);
-        line-height: 25.5px;
-        color: black;
-      }
-    }
-  }
-  >p{
-      font-size: 20px;
-      font-weight: 500;
-      color: black;
-  }
-`;
-
-const Button = styled.button`
-padding: 10px 40px;
-margin-bottom: 28px;
-font-size: 20px;
-font-weight: 600;
-letter-spacing: 1.2px;
-color: white;
-background-color: rgba(40, 65, 255, 0.8);
-border-radius: 4px;
-:hover {
-  background-color: rgba(20, 35, 205, 0.8);
-}
-:disabled {
-  background-color: rgba(0, 65, 205, 0.4);
-  cursor: not-allowed;
-}
-`;
